fix(config): satisfy pkl loader ci guards
